Horizontal Stargates and start of Crystal Tech
Wormholes are now smoother, there are two way connections, wormhole kills, Stargates can now be oriented up and down, there are now crystals which can be used inside the DHD

+++ src/main/java/net/povstalec/sgjourney/stargate/PointOfOrigin.java
@@ -69,7 +69,7 @@ public class PointOfOrigin
 		return generatesRandomly;
 	}
 	
-	public static ResourceKey<PointOfOrigin> getRandomPointOfOrigin(MinecraftServer server, int seed)
+	public static ResourceKey<PointOfOrigin> getRandomPointOfOrigin(MinecraftServer server, long seed)
 	{
 		Random random = new Random(seed);
 		RegistryAccess registries = server.registryAccess();
